change all primary key types from int to Long

+++ src/main/java/com/rmr/dinosaurs/infrastucture/database/CourseAndTagRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseAndTagRepository extends JpaRepository<CourseAndTag, Integer> {
+public interface CourseAndTagRepository extends JpaRepository<CourseAndTag, Long> {
 
 }
